% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fit_lm_poly}
\alias{fit_lm_poly}
\title{fit_lm_poly function}
\usage{
fit_lm_poly(data, vars)
}
\arguments{
\item{data}{dataset including geographic rankings classified as imd_rank}

\item{vars}{metric to create model for}
}
\value{

}
\description{
Apply linear model to supplied data based on IMD rankings.
}
\details{
Output will reflect the "expected" value based on all supplied values and IMD rank,
including expected value as well as approximate 95% confidence interval levels.

Results will be filtered to only include points every ~1000 values to support graph rendering
}
