% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fit_lm_poly_outliers}
\alias{fit_lm_poly_outliers}
\title{fit_lm_poly_outliers function}
\usage{
fit_lm_poly_outliers(data, vars)
}
\arguments{
\item{data}{dataset including geographic rankings classified as imd_rank}

\item{vars}{metric to create model for}
}
\value{

}
\description{
Summarise LSOA values, aggregated to Local Authority, based on linear model
}
\details{
Linear model will reflect expected results based on IMD rankings, identifying
any potential outliers based on approximate 95% confidence intervals
}
