% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{fit_lm_poly_vals}
\alias{fit_lm_poly_vals}
\title{fit_lm_poly_vals function}
\usage{
fit_lm_poly_vals(data, vars)
}
\arguments{
\item{data}{dataset including geographic rankings classified as imd_rank}

\item{vars}{metric to create model for}
}
\value{

}
\description{
Apply linear model to supplied data based on IMD rankings
Output will reflect the "expected" value based on all supplied values and IMD rank
}
