% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\docType{data}
\name{hcoptslang}
\alias{hcoptslang}
\title{Set global parameter for comma as thousand separator}
\format{
An object of class \code{list} of length 27.
}
\usage{
hcoptslang
}
\description{
Set global parameter for comma as thousand separator
}
\keyword{datasets}
